Fixing 'Receiver::setRXWatchPin' to take an 'int' instead of a 'uint8_t'. This way, the argument can actually be negative.

+++ src/Receiver.cpp
@@ -1029,7 +1029,7 @@ void Receiver::enableIRQs() const {
 //  RX pin interrupt and ISRs
 // ---------------------------------------------------------------------------
 
-void Receiver::setRXWatchPin(uint8_t pin) {
+void Receiver::setRXWatchPin(int pin) {
   __disable_irq();
   if (pin < 0) {
     if (rxWatchPin_ >= 0) {
